Apply filter to host before instantiating
This commit allows us to pass a reference to bigquery.Client in SpellHost
rather than instantiating it right at the bat. This solves the problem when we
run:

    $ pytest -m "not bqtest"

yet the client is still instantiated thus causing an error. With this solution,
we can then prevent the hassle of needing to have the developer install the BQ
client even if there's no need to test them.

So now, instead of passing bigquery.Client() to SpellHost, we just pass a
reference to it, i.e. bigquery.Client, and just have it instantiated when
needed. So whenever you skip all bqtests, that instantiating function
(instantiate_host()) will be skipped as well and the Client will not be
created.

Now, even if you are not connected to the internet, or even if you don't have
BQ installed, the other tests will still run.

Signed-off-by: Lester James V. Miranda <lj@thinkingmachin.es>